+++ lib/discoveredDeviceInfo.d.ts
@@ -1,0 +1,5 @@
+export interface DiscoveredDeviceInfo {
+    name: string;
+    identifierUuid: string;
+    serviceUuid: string;
+}