+++ CHANGELOG.md
@@ -2,9 +2,27 @@
 
 This file tracks released versions with the changes made to this project.
 
+## Version 2.0.2
+
+### New
+
+  * We have added the compatibility of our new Bluetooth LE scanner S721
+
+
+### Improvements
+
+  * Add discovered device info class
+
+
+### Bug fixes
+
+  * N/A
+
+
+
 ## Version 1.4.3
 
 ### New
 
-- Added a CHANGELOG.
-- Added the new device types and properties to support our Bluetooth Low Energy devices (S550, S370, S320 and S721)
+* Added a CHANGELOG.
+* Added the new device types and properties to support our Bluetooth Low Energy devices (S550, S370, S320 and S721)